requirements/base.txt
django==5.1.4
djangorestframework==3.15.2
channels==4.2.0
channels-redis==4.2.1
django-redis==5.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
daphne==4.1.2